+++ ErsatzTV.FFmpeg/Capabilities/IFFmpegCapabilities.cs
@@ -4,6 +4,7 @@ namespace ErsatzTV.FFmpeg.Capabilities;
 
 public interface IFFmpegCapabilities
 {
+    bool HasHardwareAcceleration(HardwareAccelerationMode hardwareAccelerationMode);
     bool HasDecoder(string decoder);
     bool HasEncoder(string encoder);
     bool HasFilter(string filter);